add login-function

+++ app/Http/Controllers/TasksController.php
@@ -17,11 +17,20 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            return view('tasks.index', $data);
+        }else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -52,10 +61,12 @@ class TasksController extends Controller
         ]);
         
         $task = new Task;
+        $user = \Auth::user();
+        $task->user_id = $user->id;
+        $task->content = $request->content;    
         $task->status = $request->status;
-        $task->content = $request->content;
         $task->save();
-
+        //return "hello";
         return redirect('/');
     }
 
@@ -67,11 +78,16 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task = Task::find($id);
-
+    $user = \Auth::user();
+    $tasks = Task::where([['user_id', '=', $user->id],
+                          ['id', '=', $id]]);
+    if($tasks->exists()) {
         return view('tasks.show', [
-            'task' => $task,
+            'task' => $tasks->first(),
         ]);
+    } else {
+        return redirect('/');
+    }
     }
 
     /**
@@ -119,9 +135,13 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        
+        $task = \App\Task::find($id);
 
-        return redirect('/');
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
+        }
+
+        return redirect()->back();
     }
 }
